+++ CHANGELOG.md
@@ -13,6 +13,12 @@ This project does not follow SemVer, since modules are independent of each other
 ### HTTP Proxy
 - New module to launch a tiny-proxy based HTTP proxy instance to allow an application server to assume a static IP address for outgoing traffic, e.g. when using a third-party API that requires a static IP address. For a robust and scaleable solution, consider using a NAT Gateway instead.
 
+### ecs
+- Add service_discovery_enabled as variables. [#240](https://github.com/dbl-works/terraform/pull/240)
+
+### stack/app
+- Add service_discovery_enabled as variables. [#240](https://github.com/dbl-works/terraform/pull/240)
+
 ### iam/iam-for-deploy-bot
 - add permissions to manage service discovery [#241](https://github.com/dbl-works/terraform/pull/241)